backend/requirements.txt
alembic==1.16.5
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Mail==0.10.0
Flask-Session==0.8.0
python-dotenv==1.0.0
psycopg2-binary==2.9.11
gunicorn==21.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
greenlet==3.2.4
idna==3.10
requests==2.31.0
SQLAlchemy==2.0.43
tomli==2.2.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
cachelib==0.13.0
msgspec==0.20.0
